Implement basic T&C feature (#1)
This is a preliminary implementation just to get something up and
running. It's intended to be cleaned up and generally improved in
followup PRs. Notable missing/lacking work is testing and error
handling.

The app checks the current T&C version from Wallet Proxy (WP) and asks
the user to accept it. Once accepted, the T&C version is persisted into
shared preferences. The time of the latest comparison of this value
against the one fetched from WP is stored in transient state which is
reset when the app is launched.

The current behavior is to verify T&C acceptance if it's been more than
1 minute since it happened last. And it only happens when rendering the
"home" screen. It still remains to be spec'd out completely what the
behavior actually should be.

Once the timestamp is updated, the home screen re-renders automatically
(by virtue of listening to the app state object) and replaces the T&C
widget with a dump of the stored timestamp and version along with
buttons for resetting the data for debugging purposes. Whether or not
this should be using navigation rather than simple conditional widget
replacement still isn't clear.

Added libraries:
- [`provider`](https://pub.dev/packages/provider): State management
- [`json_serializable`](https://pub.dev/packages/json_serializable): For
generating JSON models.
- [`shared_preferences`](https://pub.dev/packages/shared_preferences):
Platform-agnostic shared preferences.
- [`flutter_svg`](https://pub.dev/packages/flutter_svg): Rendering SVG
assets (the graphics on the T&C screen).