Use the bloc state management library
This enables us to move business logic out of the widgets, removing 'RefreshTermsAndConditionsScreen' entirely.

Some refactorings are in order, still. The logic of reloading the currently valid T&C version is removed as it isn't clear how it actually should be implemented. Probably we should do the check in a listener for a bloc/cubit of timestamps. But it probably should be only when we're actively navigation to the home screen (which is how it currently works).