Separate business logic from UI widgets
Ensure that all business logic is performed in "pure" non-widget code. This improves code structure, testability etc.

The widget `RefreshTermsAndConditionsScreen` is entirely removed as refresh logic doesn't belong in the view layer.

All global (shared preferences) and network-specific (wallet-proxy) services are exposed as a "service repository" provided to the root of the widget tree along with separate watchable state components ("change notifiers") for selected network and T&C acceptance. The file state.dart is split into one file in `state/` for each of these state components.

The currently valid T&C version is fetched when initializing the "home" screen. The time-based logic of reloading this version is removed as it isn't clear how it actually should be implemented. It probably should be only when we're opening/waking up the app.