iOS: Add/update project files (#13)
The previous PR was not built on a Mac and therefore didn't include the
updated iOS project files.

This change is the result of running `flutter pub get` on a Mac and
ensures that the files are in sync.

Going forward, we'll need a CI workflow to check that the project files
are in sync on all platforms (i.e. that `flutter pub get` doesn't change
any files). This may be one that triggers only on changes to
`pubspec.yaml` of the one that runs tests (see task
https://concordium.atlassian.net/browse/CBW-1482).