Keep Future of fetching valid T&C as widget state
Display the spinner while the data is being fetched, even if there is already a previously fetched version.

It isn't clear that this is the desired behavior (the user probably doesn't care that this data is being fetched in the background). We're trying it out becuase a reviewer pointed it out in a previous PR.