Separate business logic from UI widgets (#12)
## Purpose

Ensure that all business logic is performed in "pure" non-widget code.
This simplifies the widget code, structure and testability of state,
etc.

## Changes

All global (shared preferences) and network-specific (wallet-proxy)
services are exposed as a "service repository" provided to the root of
the widget tree along with separate watchable state components ("change
notifiers") for selected network and T&C acceptance. The file
`state.dart` is split into one file in `state/` for each of these state
components.

To ensure that the widgets have no dependencies to concrete networks,
the domain model is extended a little around network management. The
(currently hardcoded) selected network is injected into the widget tree
where it's used to pick the correct networked service from the service
repository.

The widget `RefreshTermsAndConditionsScreen` is entirely removed as T&C
refresh logic doesn't belong in the view builder. Instead, the currently
valid T&C version is fetched when initializing the "home" screen, which
now has to be stateful to support one-time initialization. The
time-based logic of reloading this version is removed as it isn't clear
how it actually should be implemented. It probably should be only when
we're opening/waking up the app. Regardless, all it takes to refresh is
fetching the updated data and pass it to the T&C acceptance state which
is fetched from the widget tree. The UI updates reactively to reflect
the updated state.

One outcome of this change is that the application reacts immediately
after the accepted T&C version differs from the valid one - not only
once the check time is reset as well. This check time has been changed
to the valid version fetch time, but no longer used (as mentioned
above). It also now is possible to revoke acceptance.