Refactoring: Improve initialization and network structure (#15)
## Purpose

- Prepare for initialization to be asynchronous and for config to be
fetched rather than hardcoded.
- Clarify the intended network management architecture and simplify the
way you access services belonging to the currently selected network.

## Changes

Ensured that the types of functions/methods that we know/expect are
going to become asynchronous are made so right away in order to avoid
having to change the widget structure later on. Concretely, this
includes `ServiceRepository.activateNetwork` and config loading.

Changed "loading shared preferences" into more generic "bootstrap"
procedure that is also going to load configuration and whatever other
services are going to be added (like DB client).

Changed `ActiveNetwork` to `SelectedNetwork` as, during transition, a
network can be active without being selected for a brief time. Its state
is now the actual services to access instead of just the lookup key.
This makes it simpler and less error prone to access them.
`NetworkServices` now has a reference to `Network` so this leads to no
loss of information. The only difference is that `NetworkServices` is
initialized async, meaning that the value isn't immediately available.
This is probably a good thing as it prevents you from attempting to
interact with a network before it's ready.

---------

Co-authored-by: Søren Schwartz <132270889+schwartz-concordium@users.noreply.github.com>