Rename AcceptedTermsAndConditions types
I think it's misleading to name the `AcceptedTermsAndConditionsState` "state" as it's really just a value that can be passed around, not necessarily in the context of state. It was renamed in a previous PR to give the new Hive entity its name instead. I think it's important to not refer to entity objects directly from the widgets, and therefore useful to explicitly suffix entity types with `Entity` to make it clear if it happens.

As such, the changes in this PR is just the renamings

- `AcceptedTermsAndConditionsState` -> `AcceptedTermsAndConditions`
- `AcceptedTermsAndConditions` -> `AcceptedTermsAndConditionsEntity`

Also `acceptNow` was renamed to `acceptedNow` to avoid indicating that it has a side effect.