Initial implementation of screen for configuring authentication
The password is persisted using the library `flutter_secure_storage`. There's not yet special handling of web.

A basic global authentication service `AuthenticationService` has been added. The service implements login and password update/reset. Besides this service, the authentication status is available to the app in the cubit `Authentication` which is updated in response to (async) calls to `AuthenticationService` in accordance with the usual bloc patterns.

The screen is available from a button on the home screen.