Secret storage and password set- and unlock (#29)
## Purpose

This PR continues the great work done in PR
https://github.com/Concordium/concordium-wallet/pull/25.

This PR only focus on the usage of secret storage and password set and
unlock. Implementations are done both for mobile and web.

**Web**


![password_web](https://github.com/Concordium/concordium-wallet/assets/132270889/c212dd89-62f1-44fc-924d-7f0c03139af0)

We are using Hive encrypted storage. The encryption key for the
encrypted storage is encrypted using AES GCM 256 and the users password.
The encryption is stored in a non encrypted Hive box.

- Set password
- A secure key is generated which is used to encrypt the encrypted Hive
storage.
- The users password is hashed to ensure the length is correct for the
AES GCM 256 algorithm.
- The secure key is encrypted with the users password and stored in a
non encrypted storage together with salt of the encryption.

- Unlock
    - The encrypted secure key is fetched from the non encrypted storage
- Hash of the input password is generated and used to decrypt the secure
key. If the decryption succeed the password is correct and the encrypted
Hive storage is opened / created.

- Set, read or delete from secret storage
- It is validated that the encrypted storage and a password exists. The
logic here could be moved to the business layer to ensure the user is
logged in.

**Mobile**


![password_mobile](https://github.com/Concordium/concordium-wallet/assets/132270889/f513c9c4-c794-407e-b899-706fb253f796)

When using mobile we use
[flutter_secure_storage](https://pub.dev/packages/flutter_secure_storage)
which gives encrypted storage (Keychain for iOS and
EncryptedSharedPreferences for Android).

- Set password
- Hash of the password is created using Argon2id
(https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html)

- Unlock
- Hash of the input password is created and validated against stored
hash

- Set, read or delete from secret storage
- It validates password exists. The logic here could be moved to the
business layer to ensure the user is logged in.


## Changes

-  Add logic for set and unlock using password for web and mobile
- Add logic to set, read and delete items to encrypted storages on we
and mobile

## Checklist

- [x] My code follows the style of this project.
- [x] The code compiles without warnings.
- [x] I have performed a self-review of the changes.
- [x] I have documented my code, in particular the intent of the
      hard-to-understand areas.

---------

Co-authored-by: rh-concordium <152864167+rh-concordium@users.noreply.github.com>